requirements.txt
aiogram==3.0.0b7
apprise==1.4.0
APScheduler==3.10.1
gspread==5.7.1
langchain==0.0.190
openai==0.27.8
pandas==1.5.3
psycopg2==2.9.3
pygsheets==2.0.6
python-dotenv==1.0.0
tiktoken==0.3.3
